+++ chaosslack/notification.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import json
 from typing import Any, Dict
 
 from chaoslib.types import EventPayload
@@ -214,9 +215,15 @@ def notify(settings: Dict[str, Any], event: EventPayload):
         activities_attachment.pop("footer_icon", None)
         attachments.append(rollbacks_attachment)
 
-    sc.api_call(
+    result = sc.api_call(
         "chat.postMessage",
         channel=channel,
         text="New Chaos Experiment Event",
         attachments=attachments
     )
+
+    if result.get("ok", False) is False:
+        logger.error("Slack client call failed")
+
+    logger.debug(
+        "Slack client return call: {}".format(json.dumps(result, indent=2)))