+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-slack/compare/0.1.0...HEAD
 
+### Changed
+
+-   Logging Slack Client output for debug purpose
+
 ## [0.1.0][]
 
 [0.1.0]: https://github.com/chaostoolkit-incubator/chaostoolkit-slack/tree/0.1.0